fix: cook homeboy
